# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Airline Booking System API
  version: 1.1.0
  description: |
    Contract for the Airline Booking System gateway (http://localhost:8080). All calls go through
    the API Gateway; it validates JWTs, enforces ADMIN roles and routes to the micro-services.
    Protected endpoints require `Authorization: Bearer <jwt>`. The gateway injects the
    `X-User-Email` and `X-User-Role` headers for the services, so clients do not send them.
    All errors use a uniform body: {"timestamp","status","error","message","path"}.
servers:
  - url: http://localhost:8080
    description: API Gateway (the only backend URL a frontend should call)
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register a passenger account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Account created; access and refresh tokens issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Authenticate with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Tokens issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Exchange a refresh token for a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshTokenRequest' }
      responses:
        '200':
          description: Tokens renewed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/auth/promote/{userId}:
    post:
      tags: [Auth]
      summary: Promote an existing user to ADMIN
      description: ADMIN only - enforced by the gateway (`X-User-Role` must be ADMIN).
      parameters:
        - { name: userId, in: path, required: true, schema: { type: integer, format: int64 }, description: User identifier }
      responses:
        '200':
          description: User promoted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: User promoted to ADMIN }
                  userId: { type: integer, format: int64 }
                  email: { type: string, format: email }
                  role: { type: string, example: ADMIN }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/auth/forgot-password:
    post:
      tags: [Auth]
      security: []
      summary: Request a password-reset email (no account disclosure)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmailRequest' }
      responses:
        '202':
          description: Accepted - reset email queued if the account exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: If the account exists, a reset email will be sent }
  /api/auth/reset-password:
    post:
      tags: [Auth]
      security: []
      summary: Set a new password with the emailed token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordRequest' }
      responses:
        '204': { description: Password updated }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/auth/verification:
    post:
      tags: [Auth]
      security: []
      summary: Request the account-verification email
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmailRequest' }
      responses:
        '202':
          description: Accepted - verification email requested
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: Verification email requested }
  /api/auth/verification/confirm:
    post:
      tags: [Auth]
      security: []
      summary: Verify the account with the emailed token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerificationRequest' }
      responses:
        '204': { description: Account verified }
        '400': { $ref: '#/components/responses/BadRequest' }
  # ----------------------------- Flight Service -----------------------------
...
  /api/bookings:
    get:
      tags: [Bookings]
      summary: List the current user's bookings
      responses:
        '200':
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Booking' }
    post:
      tags: [Bookings]
      summary: Create a pending booking and acquire a temporary seat lock
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BookingRequest' }
      responses:
        '201':
          description: Pending booking created (status PENDING_PAYMENT)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/SeatUnavailable' }
  /api/bookings/group:
    post:
      tags: [Bookings]
      summary: Create one pending booking per traveller (independent seat locks)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GroupBookingRequest' }
      responses:
        '201':
          description: Bookings created
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Booking' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/SeatUnavailable' }
  /api/bookings/{id}:
    get:
      tags: [Bookings]
      summary: Retrieve a single booking owned by the caller
      parameters:
        - { $ref: '#/components/parameters/BookingId' }
      responses:
        '200':
          description: Booking
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/bookings/{id}/upgrade:
    post:
      tags: [Bookings]
      summary: Change a pending booking's seat and add the price difference
      parameters:
        - { $ref: '#/components/parameters/BookingId' }
        - { name: seatNumber, in: query, required: true, schema: { type: string }, description: New seat (e.g. 1A) }
        - { name: additionalAmount, in: query, required: true, schema: { type: number, format: double, minimum: 0 } }
      responses:
        '200':
          description: Upgraded booking
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/bookings/{id}/cancel:
    post:
      tags: [Bookings]
      summary: Cancel a pending booking (releases the seat lock)
      parameters:
        - { $ref: '#/components/parameters/BookingId' }
      responses:
        '200':
          description: Booking cancelled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  # ----------------------------- Payment Service -----------------------------
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    BookingId: { name: id, in: path, required: true, schema: { type: integer, format: int64 }, description: Booking identifier }
...
  schemas:
    Error:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        path: { type: string }
    Role:
      type: string
      enum: [PASSENGER, STAFF, ADMIN]
    RegisterRequest:
      type: object
      required: [firstName, lastName, email, password]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    EmailRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token: { type: string }
        password: { type: string, minLength: 8 }
    VerificationRequest:
      type: object
      required: [token]
      properties:
        token: { type: string }
    AuthResponse:
      type: object
      properties:
        token: { type: string, description: Access token (JWT) }
        tokenType: { type: string, example: Bearer }
        userId: { type: integer, format: int64 }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
        refreshToken: { type: string }
...
    BookingStatus:
      type: string
      enum: [PENDING_PAYMENT, CONFIRMED, CANCELLED, EXPIRED]
    Booking:
      type: object
      properties:
        id: { type: integer, format: int64 }
        pnr: { type: string, description: Six-character reservation code }
        ownerEmail: { type: string, format: email }
        flightId: { type: integer, format: int64 }
        passengerId: { type: integer, format: int64 }
        seatNumber: { type: string }
        amount: { type: number, format: double }
        status: { $ref: '#/components/schemas/BookingStatus' }
        createdAt: { type: string, format: date-time }
    BookingRequest:
      type: object
      required: [flightId, passengerId, seatNumber, amount]
      properties:
        flightId: { type: integer, format: int64, minimum: 1 }
        passengerId: { type: integer, format: int64, minimum: 1 }
        seatNumber: { type: string, pattern: '^[0-9]{1,3}[A-Za-z]$' }
        amount: { type: number, format: double, exclusiveMinimum: 0 }
    TravelerSeat:
      type: object
      required: [passengerId, seatNumber, amount]
      properties:
        passengerId: { type: integer, format: int64, minimum: 1 }
        seatNumber: { type: string, pattern: '^[0-9]{1,3}[A-Za-z]$' }
        amount: { type: number, format: double, exclusiveMinimum: 0 }
    GroupBookingRequest:
      type: object
      required: [flightId, travelers]
      properties:
        flightId: { type: integer, format: int64, minimum: 1 }
        travelers:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/TravelerSeat' }
...
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: The authenticated user lacks the ADMIN role for this route
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Conflict (e.g. seat already locked, or operation not allowed in the current status)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    SeatUnavailable:
      description: The requested seat is locked by another booking attempt
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }